Refactor authentication endpoints
