+++ src/application/Endpoints/Auth/IdentityEndpoints.cs
@@ -1,10 +1,10 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Routing;
-using Super.Paula.Application.Administration.Requests;
+using Super.Paula.Application.Auth.Requests;
 using System;
 
-namespace Super.Paula.Application.Administration
+namespace Super.Paula.Application.Auth
 {
     public static class IdentityEndpoints
     {
@@ -19,41 +19,32 @@ namespace Super.Paula.Application.Administration
                 Replace,
                 Delete);
 
-            endpoints.MapCommands(
-                "/identities/{identity}",
-                ("/reset", Reset));
-
             return endpoints;
         }
 
         private static Delegate Get =>
             [Authorize("Maintainance")]
-            (IIdentityHandler handler, string identity)
+        (IIdentityHandler handler, string identity)
                 => handler.GetAsync(identity);
 
         private static Delegate GetAll =>
             [Authorize("Maintainance")]
-            (IIdentityHandler handler)
+        (IIdentityHandler handler)
                 => handler.GetAll();
 
         private static Delegate Create =>
             [Authorize("Maintainance")]
-            (IIdentityHandler handler, IdentityRequest request)
+        (IIdentityHandler handler, IdentityRequest request)
                 => handler.CreateAsync(request);
 
         private static Delegate Replace =>
             [Authorize("Maintainance")]
-            (IIdentityHandler handler, string identity, IdentityRequest request)
+        (IIdentityHandler handler, string identity, IdentityRequest request)
                 => handler.ReplaceAsync(identity, request);
 
         private static Delegate Delete =>
             [Authorize("Maintainance")]
-            (IIdentityHandler handler, string identity, [FromHeader(Name = "If-Match")] string etag)
+        (IIdentityHandler handler, string identity, [FromHeader(Name = "If-Match")] string etag)
                 => handler.DeleteAsync(identity, etag);
-
-        private static Delegate Reset =>
-            [Authorize("Maintainance")]
-            (IIdentityHandler handler, string identity, [FromHeader(Name = "If-Match")] string etag)
-                => handler.ResetAsync(identity, etag);
     }
 }
