--- src/application/Transport.Contract/Administration/Requests/SignInRequest.cs
@@ -1,18 +1,0 @@
-﻿using Super.Paula.Validation;
-using System.ComponentModel.DataAnnotations;
-
-namespace Super.Paula.Application.Administration.Requests
-{
-    public class SignInRequest
-    {
-        [Required]
-        [KebabCase]
-        [StringLength(140)]
-        [UniqueName]
-        public string Identity { get; set; } = string.Empty;
-
-        [Required]
-        [StringLength(140)]
-        public string Secret { get; set; } = string.Empty;
-    }
-}
