+++ src/application/Transport.Contract/Auth/Requests/RegisterIdentityRequest.cs
@@ -1,12 +1,10 @@
 ﻿using Super.Paula.Validation;
 using System.ComponentModel.DataAnnotations;
 
-namespace Super.Paula.Application.Administration.Requests
+namespace Super.Paula.Application.Auth.Requests
 {
-    public class IdentityRequest
+    public class RegisterIdentityRequest
     {
-        public string ETag { get; set; } = string.Empty;
-
         [Required]
         [KebabCase]
         [StringLength(140)]
@@ -16,5 +14,9 @@ namespace Super.Paula.Application.Administration.Requests
         [EmailAddress]
         [StringLength(140)]
         public string MailAddress { get; set; } = string.Empty;
+
+        [Required]
+        [StringLength(140)]
+        public string Secret { get; set; } = string.Empty;
     }
 }
