+++ src/application/Transport.Contract/Auth/Requests/SignInIdentityRequest.cs
@@ -1,0 +1,12 @@
+﻿using Super.Paula.Validation;
+using System.ComponentModel.DataAnnotations;
+
+namespace Super.Paula.Application.Auth.Requests
+{
+    public class SignInIdentityRequest
+    {
+        [Required]
+        [StringLength(140)]
+        public string Secret { get; set; } = string.Empty;
+    }
+}
