+++ src/blazor/BlazorApp/Components/Administration/SignOutButton.razor
@@ -1,6 +1,4 @@
-﻿@using Super.Paula.Application.Administration.Exceptions
-
-@inject ITranslator Translator
+﻿@inject ITranslator Translator
 @inject IAuthenticationHandler AuthenticationHandler
 @inject NavigationManager NavigationManager
