+++ src/blazor/BlazorApp/Pages/BusinessObjectInspections/Schedule.razor
