+++ src/blazor/BlazorApp/Pages/BusinessObjectInspections/View.razor
