+++ src/blazor/BlazorApp/Pages/ChangeSecret.razor
@@ -43,7 +43,7 @@
 </EditForm>
 
 @code {
-    private ChangeSecretRequest model = new ChangeSecretRequest();
+    private ChangeIdentitySecretRequest model = new ChangeIdentitySecretRequest();
 
     private async Task OnValidSubmitAsync()
     {
