+++ src/blazor/BlazorApp/Pages/Register.razor
@@ -47,7 +47,7 @@
 </EditForm>
 
 @code {
-    private RegisterRequest model = new RegisterRequest();
+    private RegisterIdentityRequest model = new RegisterIdentityRequest();
 
     private async Task OnValidSubmitAsync()
     {
