+++ src/blazor/BlazorApp/Shared/MainLayout.razor
@@ -1,7 +1,7 @@
 ﻿@inherits LayoutComponentBase
 
 @inject Application.Administration.IAuthorizationHandler AuthorizationHandler
-@inject Application.Administration.IAuthenticationHandler AuthenticationHandler
+@inject             IAuthenticationHandler AuthenticationHandler
 
 @inject NavigationManager NavigationManager   
 @inject AppEnvironment AppEnvironment
