+++ src/blazor/BlazorApp/_Imports.razor
@@ -1,4 +1,5 @@
-﻿@using System.Globalization
+﻿@using System.ComponentModel.DataAnnotations
+@using System.Globalization
 @using System.Net
 @using System.Net.Http
 @using System.Net.Http.Json
@@ -20,6 +21,9 @@
 @using Super.Paula.Application.Auditing
 @using Super.Paula.Application.Auditing.Requests
 @using Super.Paula.Application.Auditing.Responses
+@using Super.Paula.Application.Auth
+@using Super.Paula.Application.Auth.Requests
+@using Super.Paula.Application.Auth.Responses
 @using Super.Paula.Application.Guidelines
 @using Super.Paula.Application.Guidelines.Requests
 @using Super.Paula.Application.Guidelines.Responses
@@ -44,6 +48,7 @@
 @using Super.Paula.Client.Shared
 @using Super.Paula.Client.Storage
 @using Super.Paula.ErrorHandling
+@using Super.Paula.Validation
