+++ src/blazor/Client/Administration/StoredTokenAuthenticationHandler.cs
@@ -1,6 +1,7 @@
-﻿using Super.Paula.Application.Administration;
-using Super.Paula.Application.Administration.Exceptions;
-using Super.Paula.Application.Administration.Requests;
+﻿using Super.Paula.Application.Auth;
+using Super.Paula.Application.Auth.Exceptions;
+using Super.Paula.Application.Auth.Requests;
+using Super.Paula.Application.Auth.Responses;
 using Super.Paula.Authorization;
 using Super.Paula.Client.Storage;
 using System.Threading.Tasks;
@@ -20,15 +21,18 @@ namespace Super.Paula.Client.Administration
             _localStorage = localStorage;
         }
 
-        public ValueTask ChangeSecretAsync(ChangeSecretRequest request)
+        public ValueTask ChangeSecretAsync(ChangeIdentitySecretRequest request)
             => _authenticationHandler.ChangeSecretAsync(request);
 
-        public ValueTask RegisterAsync(RegisterRequest request)
+        public ValueTask RegisterAsync(RegisterIdentityRequest request)
             => _authenticationHandler.RegisterAsync(request);
 
-        public async ValueTask<string> SignInAsync(SignInRequest request)
+        public ValueTask<ResetIdentityResponse> ResetAsync(string identity, string etag)
+            => _authenticationHandler.ResetAsync(identity, etag);
+
+        public async ValueTask<string> SignInAsync(string identity, SignInIdentityRequest request)
         {
-            var response = await _authenticationHandler.SignInAsync(request);
+            var response = await _authenticationHandler.SignInAsync(identity, request);
 
             var token = response.ToToken();
             await _localStorage.SetItemAsync("token", token);
