+++ src/data/Data/_Services.cs
@@ -4,6 +4,7 @@ using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.Extensions.DependencyInjection;
 using Super.Paula.Application.Administration;
 using Super.Paula.Application.Auditing;
+using Super.Paula.Application.Auth;
 using Super.Paula.Application.Communication;
 using Super.Paula.Application.Guidelines;
 using Super.Paula.Application.Inventory;
@@ -11,6 +12,7 @@ using Super.Paula.Authorization;
 using Super.Paula.Data.Mappings;
 using Super.Paula.Data.Mappings.Administration;
 using Super.Paula.Data.Mappings.Auditing;
+using Super.Paula.Data.Mappings.Auth;
 using Super.Paula.Data.Mappings.Communication;
 using Super.Paula.Data.Mappings.Guidelines;
 using Super.Paula.Data.Mappings.Inventory;
