+++ src/server/Server/_Endpoints.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Routing;
 using Super.Paula.Application.Administration;
 using Super.Paula.Application.Auditing;
+using Super.Paula.Application.Auth;
 using Super.Paula.Application.Communication;
 using Super.Paula.Application.Guidelines;
 using Super.Paula.Application.Inventory;
