+++ src/templates/Playwright.Extensions/Auth/AuthenticationExtensions.cs
